Extracts RelatedResources type to its own file
